Added mesh class and changed DX11 to use triangle lists instead of strips

+++ XR_APILAYER_NOVENDOR_steamvr_passthrough/depth_reconstruction.cpp
@@ -519,11 +519,11 @@ void DepthReconstruction::RunThread()
 
             if (stereoConfig.StereoFiltering != StereoFiltering_None)
             {
-                if (m_confidenceLeft.size().width >= m_cvImageWidth + m_maxDisparity)
+                if ((uint32_t)m_confidenceLeft.size().width >= m_cvImageWidth + m_maxDisparity)
                 {
                     m_confidenceLeft(cv::Rect(m_maxDisparity, 0, m_cvImageWidth, m_cvImageHeight)).convertTo(leftIn[1], CV_16S, 32768.0 / 255.0);
                 }
-                if (m_confidenceRight.size().width >= m_cvImageWidth + m_maxDisparity)
+                if ((uint32_t)m_confidenceRight.size().width >= m_cvImageWidth + m_maxDisparity)
                 {
                     m_confidenceRight(cv::Rect(m_maxDisparity, 0, m_cvImageWidth, m_cvImageHeight)).convertTo(rightIn[1], CV_16S, 32768.0 / 255.0);
                 }
@@ -624,11 +624,11 @@ void DepthReconstruction::RunThread()
                     m_confidenceRight = m_wlsFilterRight->getConfidenceMap();
                 }
 
-                if (m_confidenceLeft.size().width >= m_cvImageWidth + m_maxDisparity)
+                if ((uint32_t)m_confidenceLeft.size().width >= m_cvImageWidth + m_maxDisparity)
                 {
                     m_confidenceLeft(cv::Rect(m_maxDisparity, 0, m_cvImageWidth, m_cvImageHeight)).convertTo(left, CV_8U);
                 }
-                if (m_confidenceRight.size().width >= m_cvImageWidth + m_maxDisparity)
+                if ((uint32_t)m_confidenceRight.size().width >= m_cvImageWidth + m_maxDisparity)
                 {
                     m_confidenceRight(cv::Rect(m_maxDisparity, 0, m_cvImageWidth, m_cvImageHeight)).convertTo(right, CV_8U);
                 }
